implement block timestamp for contract details
